Migra los importes de Double a BigDecimal
double no representa de forma exacta la mayoria de los importes con
centavos, y el error se hacia visible al multiplicar: tres unidades de un
producto de 8.35 devolvian 25.049999999999997 en lugar de 25.05, y ese
valor viajaba tal cual en el JSON.

- Product.price, Order.total y OrderItem.unitPrice pasan a BigDecimal,
  con columnas DECIMAL(12,2)
- Los DTOs de producto, carrito y pedido acompañan el cambio
- Los totales se calculan con add y multiply, sin pasar por double
- La validacion de precio negativo usa compareTo en lugar de <
- Los tests comparan con isEqualByComparingTo, porque equals de BigDecimal
  tambien compara la escala y considera 800.0 distinto de 800.00
- README con la decision y sus implicancias

Verificado contra la aplicacion levantada: el caso que fallaba ahora
devuelve 25.05 y 7.70 exactos.

+++ src/main/java/com/uade/e_commerce/dto/cart/CartResponseDTO.java
@@ -1,5 +1,6 @@
 package com.uade.e_commerce.dto.cart;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.AllArgsConstructor;
@@ -16,5 +17,5 @@ public class CartResponseDTO {
     private Long id;
     private Long userId;
     private List<CartItemResponseDTO> items;
-    private Double total;
+    private BigDecimal total;
 }
